pyproject.toml
[project]
name = "de-zoom-2025"
version = "0.1.0"
requires-python = "^3.12,<3.14"
dependencies = [
    "google-cloud-storage",
    "google-cloud-bigquery",
    "psycopg-binary",
    "sqlalchemy",
    "pandas",
    "dbt-core (>=1.9.2,<2.0.0)",
    "dbt-bigquery (>=1.9.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]
[tool.poetry]
description = ""
authors = ["victor <vykuang92@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.group.dev]
optional = true
[tool.poetry.group.dev.dependencies]
ipykernel = "*"
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
